Parse view data when its a variable (#108)
Parse view data when it is a variable. Adjust doc parsing injection, as you do not have the data in the function call. PHPStan is needed to determine type of variable. This can not be returned as previously done. Therefor presetting the doc, was the best solution as of now.